requirements.txt
--index-url https://download.pytorch.org/whl/cu116
torch==1.12.1+cu116
torchvision==0.13.1+cu116

tqdm
pandas
timm
numpy
Pillow
prettytable
